refactor: split pkg/storage/providers/git/plugin.go by store type (Issue #450)
plugin.go had grown to 2,354 lines containing 4 store types (GitProvider,
GitClientTenantStore, GitConfigStore, GitAuditStore) and 103 methods in a
single file. The package already had precedent for per-store files (m365,
registration, tenant, rbac), but these three store types were left behind.

Additionally, gitCommit() and gitCommitMultiple() were duplicated between
GitConfigStore and GitAuditStore with 85%+ identical logic. The initializeRepo()
pattern was similarly duplicated across all three stores.

Split the file into 5 focused files following the existing package conventions:

Changes:
- Extract git_helpers.go: shared initializeGitRepo(), gitCommitFile(), and
  gitCommitFiles() package-level functions consolidating duplicated git ops
- Extract config_store.go: GitConfigStore (~812 lines) - YAML config CRUD,
  SOPS encryption, history, batch ops; now uses shared git helpers
- Extract audit_store.go: GitAuditStore (~749 lines) - JSON audit CRUD,
  querying, filtering, sorting; now uses shared git helpers
- Extract client_tenant_store.go: GitClientTenantStore + memoryStorage
  (~165 lines) - MVP in-memory client tenant + admin consent; uses shared
  initializeGitRepo()
- Trim plugin.go to GitProvider factory only (~216 lines): factory methods +
  init() registration

All 32 tests in pkg/storage/providers/git pass without modification. Public
API surface unchanged. All platforms compile (linux/amd64/arm64, darwin,
windows/amd64). Architecture compliance and lint checks pass.

Fixes #450
