+++ pkg/storage/providers/git/plugin.go
@@ -5,21 +5,10 @@
 package git
 
 import (
-	"context"
-	"crypto/sha256"
-	"encoding/hex"
-	"encoding/json"
 	"fmt"
-	"os"
 	"os/exec"
-	"path/filepath"
-	"strings"
-	"sync"
 	"time"
 
-	"gopkg.in/yaml.v3"
-
-	"github.com/cfgis/cfgms/features/config/git"
 	"github.com/cfgis/cfgms/pkg/cache"
 	"github.com/cfgis/cfgms/pkg/storage/interfaces"
 )
@@ -60,17 +49,14 @@ func (p *GitProvider) GetCapabilities() interfaces.ProviderCapabilities {
 
 // Available checks if git is available and accessible
 func (p *GitProvider) Available() (bool, error) {
-	// Check if git is installed
 	if _, err := exec.LookPath("git"); err != nil {
 		return false, fmt.Errorf("git not found in PATH")
 	}
-
 	return true, nil
 }
 
 // CreateClientTenantStore creates a git-based client tenant store
 func (p *GitProvider) CreateClientTenantStore(config map[string]interface{}) (interfaces.ClientTenantStore, error) {
-	// Get repository path from config, or use temp directory for testing
 	repoPathStr := "/tmp/cfgms-git-test"
 	if repoPath, ok := config["repository_path"]; ok {
 		if pathStr, ok := repoPath.(string); ok && pathStr != "" {
@@ -78,7 +64,6 @@ func (p *GitProvider) CreateClientTenantStore(config map[string]interface{}) (in
 		}
 	}
 
-	// Optional remote URL for distributed deployments (MVP: not implemented)
 	remoteURL := ""
 	if remote, ok := config["remote_url"]; ok {
 		if remoteStr, ok := remote.(string); ok {
@@ -86,7 +71,6 @@ func (p *GitProvider) CreateClientTenantStore(config map[string]interface{}) (in
 		}
 	}
 
-	// Create the git store
 	store, err := NewGitClientTenantStore(repoPathStr, remoteURL)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create git client tenant store: %w", err)
@@ -97,7 +81,6 @@ func (p *GitProvider) CreateClientTenantStore(config map[string]interface{}) (in
 
 // CreateConfigStore creates a git-based configuration store
 func (p *GitProvider) CreateConfigStore(config map[string]interface{}) (interfaces.ConfigStore, error) {
-	// Get repository path from config
 	repoPathStr := "/tmp/cfgms-git-config"
 	if repoPath, ok := config["repository_path"]; ok {
 		if pathStr, ok := repoPath.(string); ok && pathStr != "" {
@@ -105,7 +88,6 @@ func (p *GitProvider) CreateConfigStore(config map[string]interface{}) (interfac
 		}
 	}
 
-	// Optional remote URL for distributed deployments
 	remoteURL := ""
 	if remote, ok := config["remote_url"]; ok {
 		if remoteStr, ok := remote.(string); ok {
@@ -113,7 +95,6 @@ func (p *GitProvider) CreateConfigStore(config map[string]interface{}) (interfac
 		}
 	}
 
-	// Create the git config store
 	store, err := NewGitConfigStore(repoPathStr, remoteURL)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create git config store: %w", err)
@@ -124,7 +105,6 @@ func (p *GitProvider) CreateConfigStore(config map[string]interface{}) (interfac
 
 // CreateAuditStore creates a git-based audit store
 func (p *GitProvider) CreateAuditStore(config map[string]interface{}) (interfaces.AuditStore, error) {
-	// Get repository path from config
 	repoPathStr := "/tmp/cfgms-git-audit"
 	if repoPath, ok := config["repository_path"]; ok {
 		if pathStr, ok := repoPath.(string); ok && pathStr != "" {
@@ -132,7 +112,6 @@ func (p *GitProvider) CreateAuditStore(config map[string]interface{}) (interface
 		}
 	}
 
-	// Optional remote URL for distributed deployments
 	remoteURL := ""
 	if remote, ok := config["remote_url"]; ok {
 		if remoteStr, ok := remote.(string); ok {
@@ -140,7 +119,6 @@ func (p *GitProvider) CreateAuditStore(config map[string]interface{}) (interface
 		}
 	}
 
-	// Create the git audit store
 	store, err := NewGitAuditStore(repoPathStr, remoteURL)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create git audit store: %w", err)
@@ -150,23 +128,20 @@ func (p *GitProvider) CreateAuditStore(config map[string]interface{}) (interface
 }
 
 func (p *GitProvider) CreateRuntimeStore(config map[string]interface{}) (interfaces.RuntimeStore, error) {
-	// Git provider implements RuntimeStore using shared cache for runtime/session data
-	// Git is for durable configuration storage, runtime sessions are ephemeral by nature
-	// Epic 6 Compliance: Uses shared cache instead of duplicate implementation
-
-	// Create shared cache with git-specific configuration
+	// Git provider implements RuntimeStore using shared cache for runtime/session data.
+	// Git is for durable configuration storage; runtime sessions are ephemeral by nature.
+	// Epic 6 Compliance: Uses shared cache instead of duplicate implementation.
 	cacheConfig := cache.CacheConfig{
 		Name:            "git-runtime",
-		MaxSessions:     1000,             // Reasonable limit for git provider usage
-		MaxRuntimeItems: 2000,             // Higher limit for runtime state
-		DefaultTTL:      1 * time.Hour,    // Shorter TTL for git provider
-		CleanupInterval: 10 * time.Minute, // Less frequent cleanup
+		MaxSessions:     1000,
+		MaxRuntimeItems: 2000,
+		DefaultTTL:      1 * time.Hour,
+		CleanupInterval: 10 * time.Minute,
 	}
 	return cache.NewRuntimeCache(cacheConfig), nil
 }
 
 func (p *GitProvider) CreateRBACStore(config map[string]interface{}) (interfaces.RBACStore, error) {
-	// Get repository path from config
 	repoPathStr := "/tmp/cfgms-git-rbac"
 	if repoPath, ok := config["repository_path"]; ok {
 		if pathStr, ok := repoPath.(string); ok && pathStr != "" {
@@ -174,7 +149,6 @@ func (p *GitProvider) CreateRBACStore(config map[string]interface{}) (interfaces
 		}
 	}
 
-	// Optional remote URL for distributed deployments
 	remoteURL := ""
 	if remote, ok := config["remote_url"]; ok {
 		if remoteStr, ok := remote.(string); ok {
@@ -182,7 +156,6 @@ func (p *GitProvider) CreateRBACStore(config map[string]interface{}) (interfaces
 		}
 	}
 
-	// Create the git RBAC store
 	store, err := NewGitRBACStore(repoPathStr, remoteURL)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create git RBAC store: %w", err)
@@ -192,7 +165,6 @@ func (p *GitProvider) CreateRBACStore(config map[string]interface{}) (interfaces
 }
 
 func (p *GitProvider) CreateTenantStore(config map[string]interface{}) (interfaces.TenantStore, error) {
-	// Get repository path from config
 	repoPathStr := "/tmp/cfgms-git-tenant"
 	if repoPath, ok := config["repository_path"]; ok {
 		if pathStr, ok := repoPath.(string); ok && pathStr != "" {
@@ -200,7 +172,6 @@ func (p *GitProvider) CreateTenantStore(config map[string]interface{}) (interfac
 		}
 	}
 
-	// Optional remote URL for distributed deployments
 	remoteURL := ""
 	if remote, ok := config["remote_url"]; ok {
 		if remoteStr, ok := remote.(string); ok {
@@ -208,7 +179,6 @@ func (p *GitProvider) CreateTenantStore(config map[string]interface{}) (interfac
 		}
 	}
 
-	// Create the git tenant store
 	store, err := NewGitTenantStore(repoPathStr, remoteURL)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create git tenant store: %w", err)
@@ -218,7 +188,6 @@ func (p *GitProvider) CreateTenantStore(config map[string]interface{}) (interfac
 }
 
 func (p *GitProvider) CreateRegistrationTokenStore(config map[string]interface{}) (interfaces.RegistrationTokenStore, error) {
-	// Get repository path from config
 	repoPathStr := "/tmp/cfgms-git-registration"
 	if repoPath, ok := config["repository_path"]; ok {
 		if pathStr, ok := repoPath.(string); ok && pathStr != "" {
@@ -226,7 +195,6 @@ func (p *GitProvider) CreateRegistrationTokenStore(config map[string]interface{}
 		}
 	}
 
-	// Optional remote URL for distributed deployments
 	remoteURL := ""
 	if remote, ok := config["remote_url"]; ok {
 		if remoteStr, ok := remote.(string); ok {
@@ -234,7 +202,6 @@ func (p *GitProvider) CreateRegistrationTokenStore(config map[string]interface{}
 		}
 	}
 
-	// Create the git registration token store
 	store, err := NewGitRegistrationTokenStore(repoPathStr, remoteURL)
 	if err != nil {
 		return nil, fmt.Errorf("failed to create git registration token store: %w", err)
@@ -246,2109 +213,4 @@ func (p *GitProvider) CreateRegistrationTokenStore(config map[string]interface{}
 // Auto-register this provider (Salt-style)
 func init() {
 	interfaces.RegisterStorageProvider(&GitProvider{})
-}
-
-// GitClientTenantStore implements ClientTenantStore using git for persistence
-type GitClientTenantStore struct {
-	repoPath  string
-	remoteURL string         // MVP: not used, for future implementation
-	storage   *memoryStorage // Per-instance storage to avoid test conflicts
-}
-
-// NewGitClientTenantStore creates a new git-based client tenant store
-func NewGitClientTenantStore(repoPath, remoteURL string) (*GitClientTenantStore, error) {
-	store := &GitClientTenantStore{
-		repoPath:  repoPath,
-		remoteURL: remoteURL,
-		storage:   newMemoryStorage(), // Each instance gets its own storage
-	}
-
-	// Initialize git repository if it doesn't exist
-	if err := store.initializeRepo(); err != nil {
-		return nil, fmt.Errorf("failed to initialize git repository: %w", err)
-	}
-
-	return store, nil
-}
-
-// initializeRepo ensures the git repository exists
-func (s *GitClientTenantStore) initializeRepo() error {
-	// Check if directory exists
-	if _, err := os.Stat(s.repoPath); os.IsNotExist(err) {
-		// Create directory
-		// #nosec G301 - Git repository directories need standard permissions for git operations
-		if err := os.MkdirAll(s.repoPath, 0755); err != nil {
-			return fmt.Errorf("failed to create directory: %w", err)
-		}
-	}
-
-	// Check if it's already a git repo
-	gitDir := fmt.Sprintf("%s/.git", s.repoPath)
-	if _, err := os.Stat(gitDir); os.IsNotExist(err) {
-		// Initialize git repository
-		cmd := exec.Command("git", "init")
-		cmd.Dir = s.repoPath
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to initialize git repository: %w", err)
-		}
-
-		// Set up initial config (basic user info for commits)
-		configCmds := [][]string{
-			{"git", "config", "user.name", "CFGMS Controller"},
-			{"git", "config", "user.email", "controller@cfgms.local"},
-			{"git", "config", "init.defaultBranch", "main"},
-		}
-
-		for _, cmdArgs := range configCmds {
-			// #nosec G204 - Git repository initialization requires controlled git config commands
-			cmd := exec.Command(cmdArgs[0], cmdArgs[1:]...)
-			cmd.Dir = s.repoPath
-			// Ignore errors for initial setup
-			_ = cmd.Run()
-		}
-	}
-
-	return nil
-}
-
-// MVP Implementation Note:
-// This is a minimal implementation to unblock the current sprint.
-// Full implementation will be done in Epic 5 with proper:
-// - JSON file storage with git commits
-// - Remote repository sync
-// - Mozilla SOPS integration
-// - Proper error handling and recovery
-// - Performance optimization
-
-// For now, we'll implement just enough to satisfy the interface
-// and use a simple in-memory storage for the MVP functionality
-// This allows the MSP feature to work while we build the full implementation
-
-// MVP: Simple in-memory storage within the git provider
-// Future: implement proper git file storage with commits
-type memoryStorage struct {
-	clients  map[string]*interfaces.ClientTenant
-	requests map[string]*interfaces.AdminConsentRequest
-	mutex    sync.RWMutex
-}
-
-func newMemoryStorage() *memoryStorage {
-	return &memoryStorage{
-		clients:  make(map[string]*interfaces.ClientTenant),
-		requests: make(map[string]*interfaces.AdminConsentRequest),
-	}
-}
-
-// Interface implementations - MVP uses simple memory storage
-// Future: implement proper git file storage with commits
-
-func (s *GitClientTenantStore) StoreClientTenant(client *interfaces.ClientTenant) error {
-	s.storage.mutex.Lock()
-	defer s.storage.mutex.Unlock()
-
-	if client.ID == "" {
-		client.ID = client.TenantID // Use tenant ID as primary key
-	}
-	if client.CreatedAt.IsZero() {
-		client.CreatedAt = time.Now()
-	}
-	client.UpdatedAt = time.Now()
-
-	s.storage.clients[client.TenantID] = client
-	return nil
-}
-
-func (s *GitClientTenantStore) GetClientTenant(tenantID string) (*interfaces.ClientTenant, error) {
-	s.storage.mutex.RLock()
-	defer s.storage.mutex.RUnlock()
-
-	client, exists := s.storage.clients[tenantID]
-	if !exists {
-		return nil, fmt.Errorf("client tenant not found: %s", tenantID)
-	}
-	return client, nil
-}
-
-func (s *GitClientTenantStore) GetClientTenantByIdentifier(clientIdentifier string) (*interfaces.ClientTenant, error) {
-	s.storage.mutex.RLock()
-	defer s.storage.mutex.RUnlock()
-
-	for _, client := range s.storage.clients {
-		if client.ClientIdentifier == clientIdentifier {
-			return client, nil
-		}
-	}
-	return nil, fmt.Errorf("client tenant not found by identifier: %s", clientIdentifier)
-}
-
-func (s *GitClientTenantStore) ListClientTenants(status interfaces.ClientTenantStatus) ([]*interfaces.ClientTenant, error) {
-	s.storage.mutex.RLock()
-	defer s.storage.mutex.RUnlock()
-
-	var result []*interfaces.ClientTenant
-	for _, client := range s.storage.clients {
-		if status == "" || client.Status == status {
-			result = append(result, client)
-		}
-	}
-	return result, nil
-}
-
-func (s *GitClientTenantStore) UpdateClientTenantStatus(tenantID string, status interfaces.ClientTenantStatus) error {
-	s.storage.mutex.Lock()
-	defer s.storage.mutex.Unlock()
-
-	client, exists := s.storage.clients[tenantID]
-	if !exists {
-		return fmt.Errorf("client tenant not found: %s", tenantID)
-	}
-
-	client.Status = status
-	client.UpdatedAt = time.Now()
-	return nil
-}
-
-func (s *GitClientTenantStore) DeleteClientTenant(tenantID string) error {
-	s.storage.mutex.Lock()
-	defer s.storage.mutex.Unlock()
-
-	delete(s.storage.clients, tenantID)
-	return nil
-}
-
-func (s *GitClientTenantStore) StoreAdminConsentRequest(request *interfaces.AdminConsentRequest) error {
-	s.storage.mutex.Lock()
-	defer s.storage.mutex.Unlock()
-
-	if request.CreatedAt.IsZero() {
-		request.CreatedAt = time.Now()
-	}
-
-	s.storage.requests[request.State] = request
-	return nil
-}
-
-func (s *GitClientTenantStore) GetAdminConsentRequest(state string) (*interfaces.AdminConsentRequest, error) {
-	s.storage.mutex.RLock()
-	defer s.storage.mutex.RUnlock()
-
-	request, exists := s.storage.requests[state]
-	if !exists {
-		return nil, fmt.Errorf("admin consent request not found: %s", state)
-	}
-
-	// Check if expired
-	if time.Now().After(request.ExpiresAt) {
-		return nil, fmt.Errorf("admin consent request expired: %s", state)
-	}
-
-	return request, nil
-}
-
-func (s *GitClientTenantStore) DeleteAdminConsentRequest(state string) error {
-	s.storage.mutex.Lock()
-	defer s.storage.mutex.Unlock()
-
-	delete(s.storage.requests, state)
-	return nil
-}
-
-// GitConfigStore implements ConfigStore using git for persistence with YAML format
-type GitConfigStore struct {
-	repoPath    string
-	remoteURL   string
-	sopsConfig  *git.SOPSConfig
-	sopsManager *git.SOPSManager
-	mutex       sync.RWMutex
-}
-
-// NewGitConfigStore creates a new git-based config store
-func NewGitConfigStore(repoPath, remoteURL string) (*GitConfigStore, error) {
-	store := &GitConfigStore{
-		repoPath:    repoPath,
-		remoteURL:   remoteURL,
-		sopsManager: git.NewSOPSManager(),
-	}
-
-	// Initialize git repository if it doesn't exist
-	if err := store.initializeRepo(); err != nil {
-		return nil, fmt.Errorf("failed to initialize git repository: %w", err)
-	}
-
-	// Load SOPS configuration if available
-	if err := store.loadSOPSConfig(); err != nil {
-		// SOPS config is optional, log error but continue
-		_ = err // SOPS config errors are non-fatal
-	}
-
-	return store, nil
-}
-
-// initializeRepo ensures the git repository exists
-func (s *GitConfigStore) initializeRepo() error {
-	// Check if directory exists
-	if _, err := os.Stat(s.repoPath); os.IsNotExist(err) {
-		// Create directory
-		// #nosec G301 - Git repository directories need standard permissions for git operations
-		if err := os.MkdirAll(s.repoPath, 0755); err != nil {
-			return fmt.Errorf("failed to create directory: %w", err)
-		}
-	}
-
-	// Check if it's already a git repo
-	gitDir := filepath.Join(s.repoPath, ".git")
-	if _, err := os.Stat(gitDir); os.IsNotExist(err) {
-		// Initialize git repository
-		cmd := exec.Command("git", "init")
-		cmd.Dir = s.repoPath
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to initialize git repository: %w", err)
-		}
-
-		// Set up initial config
-		configCmds := [][]string{
-			{"git", "config", "user.name", "CFGMS Controller"},
-			{"git", "config", "user.email", "controller@cfgms.local"},
-			{"git", "config", "init.defaultBranch", "main"},
-		}
-
-		for _, cmdArgs := range configCmds {
-			// #nosec G204 - Git repository initialization requires controlled git config commands
-			cmd := exec.Command(cmdArgs[0], cmdArgs[1:]...)
-			cmd.Dir = s.repoPath
-			_ = cmd.Run() // Ignore errors for initial setup
-		}
-	}
-
-	return nil
-}
-
-// loadSOPSConfig loads SOPS configuration from the repository
-func (s *GitConfigStore) loadSOPSConfig() error {
-	// Check for .sops.yaml file in repository root
-	sopsConfigPath := filepath.Join(s.repoPath, ".sops.yaml")
-	if _, err := os.Stat(sopsConfigPath); os.IsNotExist(err) {
-		// No SOPS config found, use default configuration
-		s.sopsConfig = &git.SOPSConfig{
-			Enabled: false,
-		}
-		return nil
-	}
-
-	// Read SOPS configuration
-	// #nosec G304 - SOPS config path is controlled by repository structure
-	configData, err := os.ReadFile(sopsConfigPath)
-	if err != nil {
-		return fmt.Errorf("failed to read SOPS config: %w", err)
-	}
-
-	config := &git.SOPSConfig{}
-	if err := yaml.Unmarshal(configData, config); err != nil {
-		return fmt.Errorf("failed to parse SOPS config: %w", err)
-	}
-
-	s.sopsConfig = config
-	return nil
-}
-
-// getConfigPath returns the file path for a config entry
-func (s *GitConfigStore) getConfigPath(key *interfaces.ConfigKey) string {
-	// Create hierarchical path structure
-	// Example: configs/tenant-a/templates/firewall.yaml
-	// Example: configs/tenant-a/certificates@device123/server-cert.yaml
-
-	var fileName string
-	if key.Scope != "" {
-		fileName = fmt.Sprintf("%s@%s.yaml", key.Name, key.Scope)
-	} else {
-		fileName = fmt.Sprintf("%s.yaml", key.Name)
-	}
-
-	return filepath.Join(s.repoPath, key.TenantID, key.Namespace, fileName)
-}
-
-// StoreConfig stores a configuration entry as YAML in git
-func (s *GitConfigStore) StoreConfig(ctx context.Context, config *interfaces.ConfigEntry) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	// Validate required fields
-	if config.Key == nil {
-		return interfaces.ErrNameRequired
-	}
-	if config.Key.TenantID == "" {
-		return interfaces.ErrTenantRequired
-	}
-	if config.Key.Namespace == "" {
-		return interfaces.ErrNamespaceRequired
-	}
-	if config.Key.Name == "" {
-		return interfaces.ErrNameRequired
-	}
-
-	// Set metadata
-	now := time.Now()
-	if config.CreatedAt.IsZero() {
-		config.CreatedAt = now
-	}
-	config.UpdatedAt = now
-	config.Format = interfaces.ConfigFormatYAML
-
-	// Calculate checksum
-	hasher := sha256.New()
-	hasher.Write(config.Data)
-	config.Checksum = hex.EncodeToString(hasher.Sum(nil))
-
-	// Get file path
-	filePath := s.getConfigPath(config.Key)
-
-	// Ensure directory exists
-	// #nosec G301 - Git repository directories need standard permissions for git operations
-	if err := os.MkdirAll(filepath.Dir(filePath), 0755); err != nil {
-		return fmt.Errorf("failed to create directory: %w", err)
-	}
-
-	// Check if file exists to determine if this is create or update
-	_, err := os.Stat(filePath)
-	isUpdate := !os.IsNotExist(err)
-
-	// Read existing config to get version
-	if isUpdate {
-		existing, err := s.readConfigFile(filePath)
-		if err != nil {
-			return fmt.Errorf("failed to read existing config: %w", err)
-		}
-		config.Version = existing.Version + 1
-	} else {
-		config.Version = 1
-	}
-
-	// Write config as YAML
-	if err := s.writeConfigFile(filePath, config); err != nil {
-		return fmt.Errorf("failed to write config file: %w", err)
-	}
-
-	// Commit to git
-	commitMsg := fmt.Sprintf("Update config %s/%s/%s (v%d)",
-		config.Key.TenantID, config.Key.Namespace, config.Key.Name, config.Version)
-	if !isUpdate {
-		commitMsg = fmt.Sprintf("Add config %s/%s/%s (v%d)",
-			config.Key.TenantID, config.Key.Namespace, config.Key.Name, config.Version)
-	}
-
-	if err := s.gitCommit(filePath, commitMsg); err != nil {
-		return fmt.Errorf("failed to commit to git: %w", err)
-	}
-
-	return nil
-}
-
-// GetConfig retrieves a configuration entry from git
-func (s *GitConfigStore) GetConfig(ctx context.Context, key *interfaces.ConfigKey) (*interfaces.ConfigEntry, error) {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	filePath := s.getConfigPath(key)
-
-	// Check if file exists
-	if _, err := os.Stat(filePath); os.IsNotExist(err) {
-		return nil, interfaces.ErrConfigNotFound
-	}
-
-	// Read config from file
-	config, err := s.readConfigFile(filePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read config: %w", err)
-	}
-
-	return config, nil
-}
-
-// DeleteConfig removes a configuration entry from git
-func (s *GitConfigStore) DeleteConfig(ctx context.Context, key *interfaces.ConfigKey) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	filePath := s.getConfigPath(key)
-
-	// Check if file exists
-	if _, err := os.Stat(filePath); os.IsNotExist(err) {
-		return interfaces.ErrConfigNotFound
-	}
-
-	// Remove file
-	if err := os.Remove(filePath); err != nil {
-		return fmt.Errorf("failed to remove config file: %w", err)
-	}
-
-	// Commit deletion to git
-	commitMsg := fmt.Sprintf("Delete config %s/%s/%s",
-		key.TenantID, key.Namespace, key.Name)
-
-	if err := s.gitCommit(filePath, commitMsg); err != nil {
-		return fmt.Errorf("failed to commit deletion to git: %w", err)
-	}
-
-	return nil
-}
-
-// ListConfigs lists configuration entries matching the filter
-func (s *GitConfigStore) ListConfigs(ctx context.Context, filter *interfaces.ConfigFilter) ([]*interfaces.ConfigEntry, error) {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	var configs []*interfaces.ConfigEntry
-
-	// Walk through the repository directory structure
-	err := filepath.Walk(s.repoPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// Skip directories and non-YAML files
-		if info.IsDir() || !strings.HasSuffix(path, ".yaml") {
-			return nil
-		}
-
-		// Skip .git directory
-		if strings.Contains(path, ".git") {
-			return nil
-		}
-
-		// Parse the config file
-		config, err := s.readConfigFile(path)
-		if err != nil {
-			// Skip files that can't be parsed
-			return nil
-		}
-
-		// Apply filters
-		if filter != nil {
-			if filter.TenantID != "" && config.Key.TenantID != filter.TenantID {
-				return nil
-			}
-			if filter.Namespace != "" && config.Key.Namespace != filter.Namespace {
-				return nil
-			}
-			if len(filter.Names) > 0 {
-				found := false
-				for _, name := range filter.Names {
-					if config.Key.Name == name {
-						found = true
-						break
-					}
-				}
-				if !found {
-					return nil
-				}
-			}
-			// Add more filter conditions as needed
-		}
-
-		configs = append(configs, config)
-		return nil
-	})
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to walk repository: %w", err)
-	}
-
-	// Apply pagination if specified
-	if filter != nil && filter.Limit > 0 {
-		start := filter.Offset
-		if start > len(configs) {
-			start = len(configs)
-		}
-		end := start + filter.Limit
-		if end > len(configs) {
-			end = len(configs)
-		}
-		configs = configs[start:end]
-	}
-
-	return configs, nil
-}
-
-// GetConfigHistory gets version history for a configuration (using git log)
-func (s *GitConfigStore) GetConfigHistory(ctx context.Context, key *interfaces.ConfigKey, limit int) ([]*interfaces.ConfigEntry, error) {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	filePath := s.getConfigPath(key)
-
-	// Make path relative to repository root for git commands
-	relPath, err := filepath.Rel(s.repoPath, filePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get relative path: %w", err)
-	}
-
-	// Convert to forward slashes for git (Windows compatibility)
-	relPath = filepath.ToSlash(relPath)
-
-	// Get git log for the file
-	args := []string{"log", "--oneline", fmt.Sprintf("-%d", limit), "--", relPath}
-	// #nosec G204 - Git storage requires controlled git command execution with validated args
-	cmd := exec.Command("git", args...)
-	cmd.Dir = s.repoPath
-
-	output, err := cmd.Output()
-	if err != nil {
-		return nil, fmt.Errorf("failed to get git history: %w", err)
-	}
-
-	lines := strings.Split(strings.TrimSpace(string(output)), "\n")
-	if len(lines) == 0 || lines[0] == "" {
-		return nil, nil
-	}
-
-	var configs []*interfaces.ConfigEntry
-
-	// For each commit, get the file contents at that point
-	for _, line := range lines {
-		if line == "" {
-			continue
-		}
-
-		parts := strings.SplitN(line, " ", 2)
-		if len(parts) < 1 {
-			continue
-		}
-
-		commitHash := parts[0]
-
-		// Get file content at this commit
-		// #nosec G204 - Git history requires accessing files at specific commits
-		cmd := exec.Command("git", "show", commitHash+":"+relPath)
-		cmd.Dir = s.repoPath
-
-		output, err := cmd.Output()
-		if err != nil {
-			continue // Skip if file doesn't exist at this commit
-		}
-
-		// Parse the historical config
-		config, err := s.parseConfigYAML(output)
-		if err != nil {
-			continue // Skip if can't parse
-		}
-
-		configs = append(configs, config)
-	}
-
-	return configs, nil
-}
-
-// GetConfigVersion gets a specific version of a configuration
-func (s *GitConfigStore) GetConfigVersion(ctx context.Context, key *interfaces.ConfigKey, version int64) (*interfaces.ConfigEntry, error) {
-	// For simplicity, we'll return the current version for now
-	// A full implementation would search through git history for the specific version
-	return s.GetConfig(ctx, key)
-}
-
-// StoreConfigBatch stores multiple configurations in a single commit
-func (s *GitConfigStore) StoreConfigBatch(ctx context.Context, configs []*interfaces.ConfigEntry) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	var filePaths []string
-
-	// Store all configs
-	for _, config := range configs {
-		if err := s.storeConfigInternal(config); err != nil {
-			return fmt.Errorf("failed to store config %s: %w", config.Key.String(), err)
-		}
-		filePaths = append(filePaths, s.getConfigPath(config.Key))
-	}
-
-	// Commit all changes in a single commit
-	commitMsg := fmt.Sprintf("Batch update %d configurations", len(configs))
-	if err := s.gitCommitMultiple(filePaths, commitMsg); err != nil {
-		return fmt.Errorf("failed to commit batch: %w", err)
-	}
-
-	return nil
-}
-
-// DeleteConfigBatch deletes multiple configurations in a single commit
-func (s *GitConfigStore) DeleteConfigBatch(ctx context.Context, keys []*interfaces.ConfigKey) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	var filePaths []string
-
-	// Delete all configs
-	for _, key := range keys {
-		filePath := s.getConfigPath(key)
-		if err := os.Remove(filePath); err != nil && !os.IsNotExist(err) {
-			return fmt.Errorf("failed to delete config %s: %w", key.String(), err)
-		}
-		filePaths = append(filePaths, filePath)
-	}
-
-	// Commit all deletions in a single commit
-	commitMsg := fmt.Sprintf("Batch delete %d configurations", len(keys))
-	if err := s.gitCommitMultiple(filePaths, commitMsg); err != nil {
-		return fmt.Errorf("failed to commit batch deletion: %w", err)
-	}
-
-	return nil
-}
-
-// ResolveConfigWithInheritance resolves configuration with inheritance (not implemented yet)
-func (s *GitConfigStore) ResolveConfigWithInheritance(ctx context.Context, key *interfaces.ConfigKey) (*interfaces.ConfigEntry, error) {
-	// For now, just return the config without inheritance
-	return s.GetConfig(ctx, key)
-}
-
-// ValidateConfig validates a configuration entry
-func (s *GitConfigStore) ValidateConfig(ctx context.Context, config *interfaces.ConfigEntry) error {
-	if config.Key == nil {
-		return interfaces.ErrNameRequired
-	}
-	if config.Key.TenantID == "" {
-		return interfaces.ErrTenantRequired
-	}
-	if config.Key.Namespace == "" {
-		return interfaces.ErrNamespaceRequired
-	}
-	if config.Key.Name == "" {
-		return interfaces.ErrNameRequired
-	}
-
-	// Validate YAML format
-	var yamlData interface{}
-	if err := yaml.Unmarshal(config.Data, &yamlData); err != nil {
-		return interfaces.ErrInvalidYAML
-	}
-
-	return nil
-}
-
-// GetConfigStats returns statistics about stored configurations
-func (s *GitConfigStore) GetConfigStats(ctx context.Context) (*interfaces.ConfigStats, error) {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	stats := &interfaces.ConfigStats{
-		ConfigsByTenant:    make(map[string]int64),
-		ConfigsByFormat:    make(map[string]int64),
-		ConfigsByNamespace: make(map[string]int64),
-		LastUpdated:        time.Now(),
-	}
-
-	// Walk through all configs to collect statistics
-	err := filepath.Walk(s.repoPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// Skip directories and non-YAML files
-		if info.IsDir() || !strings.HasSuffix(path, ".yaml") {
-			return nil
-		}
-
-		// Skip .git directory
-		if strings.Contains(path, ".git") {
-			return nil
-		}
-
-		// Parse the config file
-		config, err := s.readConfigFile(path)
-		if err != nil {
-			return nil // Skip files that can't be parsed
-		}
-
-		stats.TotalConfigs++
-		stats.TotalSize += int64(len(config.Data))
-		stats.ConfigsByTenant[config.Key.TenantID]++
-		stats.ConfigsByFormat[string(config.Format)]++
-		stats.ConfigsByNamespace[config.Key.Namespace]++
-
-		// Track oldest and newest
-		if stats.OldestConfig == nil || config.CreatedAt.Before(*stats.OldestConfig) {
-			stats.OldestConfig = &config.CreatedAt
-		}
-		if stats.NewestConfig == nil || config.UpdatedAt.After(*stats.NewestConfig) {
-			stats.NewestConfig = &config.UpdatedAt
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to collect statistics: %w", err)
-	}
-
-	// Calculate average size
-	if stats.TotalConfigs > 0 {
-		stats.AverageSize = stats.TotalSize / stats.TotalConfigs
-	}
-
-	return stats, nil
-}
-
-// Helper methods
-
-// storeConfigInternal stores a config without committing (for batch operations)
-func (s *GitConfigStore) storeConfigInternal(config *interfaces.ConfigEntry) error {
-	// Set metadata
-	now := time.Now()
-	if config.CreatedAt.IsZero() {
-		config.CreatedAt = now
-	}
-	config.UpdatedAt = now
-	config.Format = interfaces.ConfigFormatYAML
-
-	// Calculate checksum
-	hasher := sha256.New()
-	hasher.Write(config.Data)
-	config.Checksum = hex.EncodeToString(hasher.Sum(nil))
-
-	// Get file path
-	filePath := s.getConfigPath(config.Key)
-
-	// Ensure directory exists
-	// #nosec G301 - Git repository directories need standard permissions for git operations
-	if err := os.MkdirAll(filepath.Dir(filePath), 0755); err != nil {
-		return fmt.Errorf("failed to create directory: %w", err)
-	}
-
-	// Check if file exists to determine version
-	_, err := os.Stat(filePath)
-	isUpdate := !os.IsNotExist(err)
-
-	if isUpdate {
-		existing, err := s.readConfigFile(filePath)
-		if err != nil {
-			return fmt.Errorf("failed to read existing config: %w", err)
-		}
-		config.Version = existing.Version + 1
-	} else {
-		config.Version = 1
-	}
-
-	// Write config as YAML
-	if err := s.writeConfigFile(filePath, config); err != nil {
-		return fmt.Errorf("failed to write config file: %w", err)
-	}
-
-	return nil
-}
-
-// writeConfigFile writes a config entry to a YAML file with optional SOPS encryption
-func (s *GitConfigStore) writeConfigFile(filePath string, config *interfaces.ConfigEntry) error {
-	// Create a structure that includes both the metadata and the actual data
-	fileData := map[string]interface{}{
-		"metadata": map[string]interface{}{
-			"tenant_id":  config.Key.TenantID,
-			"namespace":  config.Key.Namespace,
-			"name":       config.Key.Name,
-			"scope":      config.Key.Scope,
-			"version":    config.Version,
-			"checksum":   config.Checksum,
-			"format":     config.Format,
-			"created_at": config.CreatedAt,
-			"updated_at": config.UpdatedAt,
-			"created_by": config.CreatedBy,
-			"updated_by": config.UpdatedBy,
-			"tags":       config.Tags,
-			"source":     config.Source,
-		},
-	}
-
-	// Parse the config data as YAML and add it
-	var configData interface{}
-	if err := yaml.Unmarshal(config.Data, &configData); err != nil {
-		return fmt.Errorf("failed to parse config data as YAML: %w", err)
-	}
-	fileData["config"] = configData
-
-	// Marshal the complete structure to YAML
-	yamlData, err := yaml.Marshal(fileData)
-	if err != nil {
-		return fmt.Errorf("failed to marshal to YAML: %w", err)
-	}
-
-	// Apply SOPS encryption if enabled and file should be encrypted
-	finalData := yamlData
-	if s.sopsConfig != nil && s.sopsConfig.Enabled {
-		relPath, err := filepath.Rel(s.repoPath, filePath)
-		if err != nil {
-			relPath = filePath
-		}
-
-		shouldEncrypt, _ := s.sopsManager.ShouldEncryptFile(relPath, s.sopsConfig)
-		if shouldEncrypt {
-			ctx := context.Background()
-			encryptedData, err := s.sopsManager.EncryptContent(ctx, yamlData, s.sopsConfig, relPath)
-			if err != nil {
-				// Log error but continue with unencrypted data
-				// In production, you might want to handle this differently
-				_ = err
-			} else {
-				finalData = encryptedData
-			}
-		}
-	}
-
-	// Write to file
-	// #nosec G306 - Configuration files need read permissions for other processes
-	if err := os.WriteFile(filePath, finalData, 0644); err != nil {
-		return fmt.Errorf("failed to write file: %w", err)
-	}
-
-	return nil
-}
-
-// readConfigFile reads a config entry from a YAML file with optional SOPS decryption
-func (s *GitConfigStore) readConfigFile(filePath string) (*interfaces.ConfigEntry, error) {
-	// Read file contents
-	// #nosec G304 - Git storage requires reading config files from controlled repository paths
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read file: %w", err)
-	}
-
-	// Decrypt SOPS encrypted content if necessary
-	finalData := data
-	if s.sopsManager != nil && s.sopsManager.IsSOPSEncrypted(data) {
-		ctx := context.Background()
-		decryptedData, err := s.sopsManager.DecryptContent(ctx, data)
-		if err != nil {
-			return nil, fmt.Errorf("failed to decrypt SOPS content: %w", err)
-		}
-		finalData = decryptedData
-	}
-
-	return s.parseConfigYAML(finalData)
-}
-
-// parseConfigYAML parses YAML data into a ConfigEntry
-func (s *GitConfigStore) parseConfigYAML(data []byte) (*interfaces.ConfigEntry, error) {
-	// Parse the YAML structure
-	var fileData map[string]interface{}
-	if err := yaml.Unmarshal(data, &fileData); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal YAML: %w", err)
-	}
-
-	// Extract metadata
-	metadataRaw, ok := fileData["metadata"]
-	if !ok {
-		return nil, fmt.Errorf("missing metadata section")
-	}
-
-	metadata, ok := metadataRaw.(map[string]interface{})
-	if !ok {
-		return nil, fmt.Errorf("invalid metadata format")
-	}
-
-	// Create config entry
-	config := &interfaces.ConfigEntry{
-		Key: &interfaces.ConfigKey{},
-	}
-
-	// Parse metadata fields
-	if val, ok := metadata["tenant_id"].(string); ok {
-		config.Key.TenantID = val
-	}
-	if val, ok := metadata["namespace"].(string); ok {
-		config.Key.Namespace = val
-	}
-	if val, ok := metadata["name"].(string); ok {
-		config.Key.Name = val
-	}
-	if val, ok := metadata["scope"].(string); ok {
-		config.Key.Scope = val
-	}
-	if val, ok := metadata["version"].(int); ok {
-		config.Version = int64(val)
-	}
-	if val, ok := metadata["checksum"].(string); ok {
-		config.Checksum = val
-	}
-	if val, ok := metadata["format"].(string); ok {
-		config.Format = interfaces.ConfigFormat(val)
-	}
-	if val, ok := metadata["created_by"].(string); ok {
-		config.CreatedBy = val
-	}
-	if val, ok := metadata["updated_by"].(string); ok {
-		config.UpdatedBy = val
-	}
-	if val, ok := metadata["source"].(string); ok {
-		config.Source = val
-	}
-
-	// Parse time fields
-	if val, ok := metadata["created_at"].(time.Time); ok {
-		config.CreatedAt = val
-	}
-	if val, ok := metadata["updated_at"].(time.Time); ok {
-		config.UpdatedAt = val
-	}
-
-	// Parse tags
-	if val, ok := metadata["tags"].([]interface{}); ok {
-		for _, tag := range val {
-			if tagStr, ok := tag.(string); ok {
-				config.Tags = append(config.Tags, tagStr)
-			}
-		}
-	}
-
-	// Extract config data and marshal it back to YAML
-	configData, ok := fileData["config"]
-	if !ok {
-		return nil, fmt.Errorf("missing config section")
-	}
-
-	configYAML, err := yaml.Marshal(configData)
-	if err != nil {
-		return nil, fmt.Errorf("failed to marshal config data: %w", err)
-	}
-	config.Data = configYAML
-
-	return config, nil
-}
-
-// gitCommit commits a single file to git
-func (s *GitConfigStore) gitCommit(filePath, message string) error {
-	// Add file to staging
-	// #nosec G204 - Git storage requires controlled git command execution
-	cmd := exec.Command("git", "add", filePath)
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to add file to git: %w", err)
-	}
-
-	// Commit the change
-	// #nosec G204 - Git storage requires controlled git command execution
-	cmd = exec.Command("git", "commit", "-m", message)
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to commit to git: %w", err)
-	}
-
-	return nil
-}
-
-// gitCommitMultiple commits multiple files to git in a single commit
-func (s *GitConfigStore) gitCommitMultiple(filePaths []string, message string) error {
-	// Add all files to staging
-	args := append([]string{"add"}, filePaths...)
-	// #nosec G204 - Git storage requires controlled git command execution with validated args
-	cmd := exec.Command("git", args...)
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to add files to git: %w", err)
-	}
-
-	// Commit the changes
-	// #nosec G204 - Git storage requires controlled git command execution
-	cmd = exec.Command("git", "commit", "-m", message)
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to commit to git: %w", err)
-	}
-
-	return nil
-}
-
-// ConfigureSOPS configures SOPS encryption for the repository
-func (s *GitConfigStore) ConfigureSOPS(config *git.SOPSConfig) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	// Update SOPS configuration
-	s.sopsConfig = config
-
-	if config.Enabled {
-		// Validate SOPS configuration
-		ctx := context.Background()
-		if err := s.sopsManager.ValidateSOPSConfig(ctx, config); err != nil {
-			return fmt.Errorf("invalid SOPS configuration: %w", err)
-		}
-
-		// Generate .sops.yaml configuration file
-		if err := s.sopsManager.GenerateSOPSConfig(config, s.repoPath); err != nil {
-			return fmt.Errorf("failed to generate SOPS config: %w", err)
-		}
-
-		// Commit the SOPS configuration
-		sopsConfigPath := filepath.Join(s.repoPath, ".sops.yaml")
-		if err := s.gitCommit(sopsConfigPath, "Configure SOPS encryption"); err != nil {
-			return fmt.Errorf("failed to commit SOPS config: %w", err)
-		}
-	}
-
-	return nil
-}
-
-// IsSOPSEnabled returns whether SOPS encryption is enabled
-func (s *GitConfigStore) IsSOPSEnabled() bool {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	return s.sopsConfig != nil && s.sopsConfig.Enabled
-}
-
-// SyncWithRemote synchronizes the repository with the remote
-func (s *GitConfigStore) SyncWithRemote(ctx context.Context) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	if s.remoteURL == "" {
-		return nil // No remote configured, skip sync
-	}
-
-	// Pull latest changes from remote
-	if err := s.pullFromRemote(ctx); err != nil {
-		return fmt.Errorf("failed to pull from remote: %w", err)
-	}
-
-	// Push local changes to remote
-	if err := s.pushToRemote(ctx); err != nil {
-		return fmt.Errorf("failed to push to remote: %w", err)
-	}
-
-	return nil
-}
-
-// pullFromRemote pulls changes from the remote repository
-func (s *GitConfigStore) pullFromRemote(ctx context.Context) error {
-	if s.remoteURL == "" {
-		return nil
-	}
-
-	// Check if remote origin exists
-	cmd := exec.CommandContext(ctx, "git", "remote", "get-url", "origin")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		// Add remote origin if it doesn't exist
-		// #nosec G204 - Git storage requires remote URL management for distributed repos
-		cmd = exec.CommandContext(ctx, "git", "remote", "add", "origin", s.remoteURL)
-		cmd.Dir = s.repoPath
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to add remote origin: %w", err)
-		}
-	}
-
-	// Fetch from remote
-	cmd = exec.CommandContext(ctx, "git", "fetch", "origin")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		// Fetch may fail if remote is empty, which is okay for new repos
-		return nil
-	}
-
-	// Check if we have any commits locally
-	cmd = exec.CommandContext(ctx, "git", "rev-parse", "HEAD")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		// No local commits, try to pull if remote has commits
-		cmd = exec.CommandContext(ctx, "git", "pull", "origin", "main")
-		cmd.Dir = s.repoPath
-		_ = cmd.Run() // Ignore errors as remote might be empty
-		return nil
-	}
-
-	// Pull changes
-	cmd = exec.CommandContext(ctx, "git", "pull", "origin", "main")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		// Pull may fail due to divergent histories, use rebase
-		cmd = exec.CommandContext(ctx, "git", "pull", "--rebase", "origin", "main")
-		cmd.Dir = s.repoPath
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to pull with rebase: %w", err)
-		}
-	}
-
-	return nil
-}
-
-// pushToRemote pushes changes to the remote repository
-func (s *GitConfigStore) pushToRemote(ctx context.Context) error {
-	if s.remoteURL == "" {
-		return nil
-	}
-
-	// Check if we have any local commits
-	cmd := exec.CommandContext(ctx, "git", "rev-parse", "HEAD")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return nil // No commits to push
-	}
-
-	// Check if we have unpushed changes
-	cmd = exec.CommandContext(ctx, "git", "status", "--porcelain")
-	cmd.Dir = s.repoPath
-	output, err := cmd.Output()
-	if err != nil {
-		return fmt.Errorf("failed to check git status: %w", err)
-	}
-
-	// If there are no changes and we're up to date, no need to push
-	if len(output) == 0 {
-		// Check if we're ahead of remote
-		cmd = exec.CommandContext(ctx, "git", "rev-list", "--count", "HEAD", "^origin/main")
-		cmd.Dir = s.repoPath
-		output, err := cmd.Output()
-		if err != nil || strings.TrimSpace(string(output)) == "0" {
-			return nil // No commits to push
-		}
-	}
-
-	// Push to remote
-	cmd = exec.CommandContext(ctx, "git", "push", "origin", "main")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to push to remote: %w", err)
-	}
-
-	return nil
-}
-
-// SetRemoteURL sets the remote repository URL
-func (s *GitConfigStore) SetRemoteURL(remoteURL string) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	s.remoteURL = remoteURL
-
-	if remoteURL != "" {
-		// Update or add remote origin
-		cmd := exec.Command("git", "remote", "set-url", "origin", remoteURL)
-		cmd.Dir = s.repoPath
-		if err := cmd.Run(); err != nil {
-			// If set-url fails, try to add the remote
-			cmd = exec.Command("git", "remote", "add", "origin", remoteURL)
-			cmd.Dir = s.repoPath
-			if err := cmd.Run(); err != nil {
-				return fmt.Errorf("failed to set remote URL: %w", err)
-			}
-		}
-	}
-
-	return nil
-}
-
-// GitAuditStore implements AuditStore using git for persistence with JSON format
-type GitAuditStore struct {
-	repoPath  string
-	remoteURL string
-	mutex     sync.RWMutex
-}
-
-// NewGitAuditStore creates a new git-based audit store
-func NewGitAuditStore(repoPath, remoteURL string) (*GitAuditStore, error) {
-	store := &GitAuditStore{
-		repoPath:  repoPath,
-		remoteURL: remoteURL,
-	}
-
-	// Initialize git repository if it doesn't exist
-	if err := store.initializeRepo(); err != nil {
-		return nil, fmt.Errorf("failed to initialize git repository: %w", err)
-	}
-
-	return store, nil
-}
-
-// initializeRepo ensures the git repository exists
-func (s *GitAuditStore) initializeRepo() error {
-	// Check if directory exists
-	if _, err := os.Stat(s.repoPath); os.IsNotExist(err) {
-		// Create directory
-		// #nosec G301 - Git repository directories need standard permissions for git operations
-		if err := os.MkdirAll(s.repoPath, 0755); err != nil {
-			return fmt.Errorf("failed to create directory: %w", err)
-		}
-	}
-
-	// Check if it's already a git repo
-	gitDir := filepath.Join(s.repoPath, ".git")
-	if _, err := os.Stat(gitDir); os.IsNotExist(err) {
-		// Initialize git repository
-		cmd := exec.Command("git", "init")
-		cmd.Dir = s.repoPath
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to initialize git repository: %w", err)
-		}
-
-		// Set up initial config
-		configCmds := [][]string{
-			{"git", "config", "user.name", "CFGMS Controller"},
-			{"git", "config", "user.email", "controller@cfgms.local"},
-			{"git", "config", "init.defaultBranch", "main"},
-		}
-
-		for _, cmdArgs := range configCmds {
-			// #nosec G204 - Git repository initialization requires controlled git config commands
-			cmd := exec.Command(cmdArgs[0], cmdArgs[1:]...)
-			cmd.Dir = s.repoPath
-			_ = cmd.Run() // Ignore errors for initial setup
-		}
-	}
-
-	return nil
-}
-
-// getAuditPath returns the file path for an audit entry
-// Uses date-based hierarchical structure for efficient organization
-func (s *GitAuditStore) getAuditPath(entry *interfaces.AuditEntry) string {
-	// Create hierarchical path by date for efficient organization
-	// Example: audit/2025/01/15/tenant-a/authentication-events.json
-	year := entry.Timestamp.Format("2006")
-	month := entry.Timestamp.Format("01")
-	day := entry.Timestamp.Format("02")
-
-	// Group by event type for efficient access patterns
-	fileName := fmt.Sprintf("%s-events.json", entry.EventType)
-
-	return filepath.Join(s.repoPath, year, month, day, entry.TenantID, fileName)
-}
-
-// StoreAuditEntry stores an audit entry as JSON in git
-func (s *GitAuditStore) StoreAuditEntry(ctx context.Context, entry *interfaces.AuditEntry) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	// Validate required fields
-	if entry.TenantID == "" {
-		return interfaces.ErrTenantIDRequired
-	}
-	if entry.UserID == "" {
-		return interfaces.ErrUserIDRequired
-	}
-	if entry.Action == "" {
-		return interfaces.ErrActionRequired
-	}
-	if entry.ResourceType == "" {
-		return interfaces.ErrResourceTypeRequired
-	}
-	if entry.ResourceID == "" {
-		return interfaces.ErrResourceIDRequired
-	}
-
-	// Set metadata
-	if entry.ID == "" {
-		entry.ID = s.generateAuditID(entry)
-	}
-	if entry.Timestamp.IsZero() {
-		entry.Timestamp = time.Now()
-	}
-
-	// The audit manager is responsible for checksum calculation
-	// Storage providers should preserve checksums, not generate them
-
-	// Get file path
-	filePath := s.getAuditPath(entry)
-
-	// Ensure directory exists
-	// #nosec G301 - Git repository directories need standard permissions for git operations
-	if err := os.MkdirAll(filepath.Dir(filePath), 0755); err != nil {
-		return fmt.Errorf("failed to create directory: %w", err)
-	}
-
-	// Read existing entries from the file if it exists
-	var entries []*interfaces.AuditEntry
-	if _, err := os.Stat(filePath); !os.IsNotExist(err) {
-		entries, err = s.readAuditFile(filePath)
-		if err != nil {
-			return fmt.Errorf("failed to read existing audit file: %w", err)
-		}
-	}
-
-	// Append new entry
-	entries = append(entries, entry)
-
-	// Write updated entries back to file
-	if err := s.writeAuditFile(filePath, entries); err != nil {
-		return fmt.Errorf("failed to write audit file: %w", err)
-	}
-
-	// Commit to git
-	commitMsg := fmt.Sprintf("Add audit entry %s: %s by %s",
-		entry.EventType, entry.Action, entry.UserID)
-
-	if err := s.gitCommit(filePath, commitMsg); err != nil {
-		return fmt.Errorf("failed to commit to git: %w", err)
-	}
-
-	return nil
-}
-
-// GetAuditEntry retrieves a specific audit entry by ID
-func (s *GitAuditStore) GetAuditEntry(ctx context.Context, id string) (*interfaces.AuditEntry, error) {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	// Since we don't know which file contains the entry, we need to search
-	// This is inefficient but follows the interface requirement
-	// A production implementation would use indexing
-
-	var foundEntry *interfaces.AuditEntry
-
-	err := filepath.Walk(s.repoPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// Skip directories and non-JSON files
-		if info.IsDir() || !strings.HasSuffix(path, ".json") {
-			return nil
-		}
-
-		// Skip .git directory
-		if strings.Contains(path, ".git") {
-			return nil
-		}
-
-		// Read and search this file
-		entries, err := s.readAuditFile(path)
-		if err != nil {
-			return nil // Skip files that can't be parsed
-		}
-
-		for _, entry := range entries {
-			if entry.ID == id {
-				foundEntry = entry
-				return filepath.SkipDir // Stop searching
-			}
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to search for audit entry: %w", err)
-	}
-
-	if foundEntry == nil {
-		return nil, interfaces.ErrAuditNotFound
-	}
-
-	return foundEntry, nil
-}
-
-// ListAuditEntries lists audit entries matching the filter
-func (s *GitAuditStore) ListAuditEntries(ctx context.Context, filter *interfaces.AuditFilter) ([]*interfaces.AuditEntry, error) {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	var allEntries []*interfaces.AuditEntry
-
-	// Walk through the repository directory structure
-	err := filepath.Walk(s.repoPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// Skip directories and non-JSON files
-		if info.IsDir() || !strings.HasSuffix(path, ".json") {
-			return nil
-		}
-
-		// Skip .git directory
-		if strings.Contains(path, ".git") {
-			return nil
-		}
-
-		// Read entries from this file
-		entries, err := s.readAuditFile(path)
-		if err != nil {
-			return nil // Skip files that can't be parsed
-		}
-
-		// Apply filters
-		for _, entry := range entries {
-			if s.matchesFilter(entry, filter) {
-				allEntries = append(allEntries, entry)
-			}
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to walk repository: %w", err)
-	}
-
-	// Sort by timestamp (descending by default)
-	if len(allEntries) > 1 {
-		s.sortAuditEntries(allEntries, filter)
-	}
-
-	// Apply pagination if specified
-	if filter != nil && filter.Limit > 0 {
-		start := filter.Offset
-		if start > len(allEntries) {
-			start = len(allEntries)
-		}
-		end := start + filter.Limit
-		if end > len(allEntries) {
-			end = len(allEntries)
-		}
-		allEntries = allEntries[start:end]
-	}
-
-	return allEntries, nil
-}
-
-// StoreAuditBatch stores multiple audit entries efficiently
-func (s *GitAuditStore) StoreAuditBatch(ctx context.Context, entries []*interfaces.AuditEntry) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	// Group entries by file path for efficient batch writes
-	fileGroups := make(map[string][]*interfaces.AuditEntry)
-
-	for _, entry := range entries {
-		// Validate and set metadata for each entry
-		if err := s.validateAndSetMetadata(entry); err != nil {
-			return fmt.Errorf("failed to validate entry %s: %w", entry.ID, err)
-		}
-
-		filePath := s.getAuditPath(entry)
-		fileGroups[filePath] = append(fileGroups[filePath], entry)
-	}
-
-	var filePaths []string
-
-	// Write entries grouped by file
-	for filePath, groupEntries := range fileGroups {
-		// Ensure directory exists
-		// #nosec G301 - Git repository directories need standard permissions for git operations
-		if err := os.MkdirAll(filepath.Dir(filePath), 0755); err != nil {
-			return fmt.Errorf("failed to create directory: %w", err)
-		}
-
-		// Read existing entries from the file if it exists
-		var existingEntries []*interfaces.AuditEntry
-		if _, err := os.Stat(filePath); !os.IsNotExist(err) {
-			existing, err := s.readAuditFile(filePath)
-			if err != nil {
-				return fmt.Errorf("failed to read existing audit file: %w", err)
-			}
-			existingEntries = existing
-		}
-
-		// Append new entries
-		allEntries := append(existingEntries, groupEntries...)
-
-		// Write updated entries back to file
-		if err := s.writeAuditFile(filePath, allEntries); err != nil {
-			return fmt.Errorf("failed to write audit file: %w", err)
-		}
-
-		filePaths = append(filePaths, filePath)
-	}
-
-	// Commit all changes in a single commit
-	commitMsg := fmt.Sprintf("Batch add %d audit entries", len(entries))
-	if err := s.gitCommitMultiple(filePaths, commitMsg); err != nil {
-		return fmt.Errorf("failed to commit batch: %w", err)
-	}
-
-	return nil
-}
-
-// GetAuditsByUser gets audit entries for a specific user
-func (s *GitAuditStore) GetAuditsByUser(ctx context.Context, userID string, timeRange *interfaces.TimeRange) ([]*interfaces.AuditEntry, error) {
-	filter := &interfaces.AuditFilter{
-		UserIDs:   []string{userID},
-		TimeRange: timeRange,
-	}
-	return s.ListAuditEntries(ctx, filter)
-}
-
-// GetAuditsByResource gets audit entries for a specific resource
-func (s *GitAuditStore) GetAuditsByResource(ctx context.Context, resourceType, resourceID string, timeRange *interfaces.TimeRange) ([]*interfaces.AuditEntry, error) {
-	filter := &interfaces.AuditFilter{
-		ResourceTypes: []string{resourceType},
-		ResourceIDs:   []string{resourceID},
-		TimeRange:     timeRange,
-	}
-	return s.ListAuditEntries(ctx, filter)
-}
-
-// GetAuditsByAction gets audit entries for a specific action
-func (s *GitAuditStore) GetAuditsByAction(ctx context.Context, action string, timeRange *interfaces.TimeRange) ([]*interfaces.AuditEntry, error) {
-	filter := &interfaces.AuditFilter{
-		Actions:   []string{action},
-		TimeRange: timeRange,
-	}
-	return s.ListAuditEntries(ctx, filter)
-}
-
-// GetFailedActions gets recent failed actions for security monitoring
-func (s *GitAuditStore) GetFailedActions(ctx context.Context, timeRange *interfaces.TimeRange, limit int) ([]*interfaces.AuditEntry, error) {
-	filter := &interfaces.AuditFilter{
-		Results:   []interfaces.AuditResult{interfaces.AuditResultFailure, interfaces.AuditResultError, interfaces.AuditResultDenied},
-		TimeRange: timeRange,
-		Limit:     limit,
-		SortBy:    "timestamp",
-		Order:     "desc",
-	}
-	return s.ListAuditEntries(ctx, filter)
-}
-
-// GetSuspiciousActivity gets suspicious activity for a tenant
-func (s *GitAuditStore) GetSuspiciousActivity(ctx context.Context, tenantID string, timeRange *interfaces.TimeRange) ([]*interfaces.AuditEntry, error) {
-	filter := &interfaces.AuditFilter{
-		TenantID:   tenantID,
-		EventTypes: []interfaces.AuditEventType{interfaces.AuditEventSecurityEvent},
-		Severities: []interfaces.AuditSeverity{interfaces.AuditSeverityHigh, interfaces.AuditSeverityCritical},
-		TimeRange:  timeRange,
-		SortBy:     "timestamp",
-		Order:      "desc",
-	}
-	return s.ListAuditEntries(ctx, filter)
-}
-
-// GetAuditStats returns statistics about stored audit entries
-func (s *GitAuditStore) GetAuditStats(ctx context.Context) (*interfaces.AuditStats, error) {
-	s.mutex.RLock()
-	defer s.mutex.RUnlock()
-
-	stats := &interfaces.AuditStats{
-		EntriesByTenant:   make(map[string]int64),
-		EntriesByType:     make(map[string]int64),
-		EntriesByResult:   make(map[string]int64),
-		EntriesBySeverity: make(map[string]int64),
-		LastUpdated:       time.Now(),
-	}
-
-	now := time.Now()
-	last24h := now.Add(-24 * time.Hour)
-	last7d := now.Add(-7 * 24 * time.Hour)
-	last30d := now.Add(-30 * 24 * time.Hour)
-
-	// Walk through all audit files to collect statistics
-	err := filepath.Walk(s.repoPath, func(path string, info os.FileInfo, err error) error {
-		if err != nil {
-			return err
-		}
-
-		// Skip directories and non-JSON files
-		if info.IsDir() || !strings.HasSuffix(path, ".json") {
-			return nil
-		}
-
-		// Skip .git directory
-		if strings.Contains(path, ".git") {
-			return nil
-		}
-
-		// Read entries from this file
-		entries, err := s.readAuditFile(path)
-		if err != nil {
-			return nil // Skip files that can't be parsed
-		}
-
-		for _, entry := range entries {
-			stats.TotalEntries++
-			stats.TotalSize += int64(len(entry.ID)) // Approximate size
-			stats.EntriesByTenant[entry.TenantID]++
-			stats.EntriesByType[string(entry.EventType)]++
-			stats.EntriesByResult[string(entry.Result)]++
-			stats.EntriesBySeverity[string(entry.Severity)]++
-
-			// Track oldest and newest
-			if stats.OldestEntry == nil || entry.Timestamp.Before(*stats.OldestEntry) {
-				stats.OldestEntry = &entry.Timestamp
-			}
-			if stats.NewestEntry == nil || entry.Timestamp.After(*stats.NewestEntry) {
-				stats.NewestEntry = &entry.Timestamp
-			}
-
-			// Time-based counts
-			if entry.Timestamp.After(last24h) {
-				stats.EntriesLast24h++
-			}
-			if entry.Timestamp.After(last7d) {
-				stats.EntriesLast7d++
-			}
-			if entry.Timestamp.After(last30d) {
-				stats.EntriesLast30d++
-			}
-
-			// Security statistics
-			if entry.Result == interfaces.AuditResultFailure || entry.Result == interfaces.AuditResultError || entry.Result == interfaces.AuditResultDenied {
-				if entry.Timestamp.After(last24h) {
-					stats.FailedActionsLast24h++
-				}
-			}
-
-			if entry.EventType == interfaces.AuditEventSecurityEvent {
-				stats.SuspiciousActivityCount++
-				if stats.LastSecurityIncident == nil || entry.Timestamp.After(*stats.LastSecurityIncident) {
-					stats.LastSecurityIncident = &entry.Timestamp
-				}
-			}
-		}
-
-		return nil
-	})
-
-	if err != nil {
-		return nil, fmt.Errorf("failed to collect statistics: %w", err)
-	}
-
-	// Calculate average size
-	if stats.TotalEntries > 0 {
-		stats.AverageSize = stats.TotalSize / stats.TotalEntries
-	}
-
-	return stats, nil
-}
-
-// ArchiveAuditEntries archives old audit entries (git-based implementation keeps all data)
-func (s *GitAuditStore) ArchiveAuditEntries(ctx context.Context, beforeDate time.Time) (int64, error) {
-	// Git-based storage naturally archives data through commits
-	// For compliance, we could implement a separate archive branch
-	// For now, return 0 as no entries are physically moved
-	return 0, nil
-}
-
-// PurgeAuditEntries purges very old audit entries (use with caution)
-func (s *GitAuditStore) PurgeAuditEntries(ctx context.Context, beforeDate time.Time) (int64, error) {
-	// Implementation would remove entries from files and commit the changes
-	// This is a destructive operation and should be used carefully
-	// For now, return 0 as no entries are purged
-	return 0, nil
-}
-
-// Helper methods for GitAuditStore
-
-// generateAuditID generates a unique ID for an audit entry
-func (s *GitAuditStore) generateAuditID(entry *interfaces.AuditEntry) string {
-	// Create a deterministic ID based on entry contents and timestamp
-	data := fmt.Sprintf("%s-%s-%s-%s-%d",
-		entry.TenantID, entry.UserID, entry.Action, entry.ResourceID, entry.Timestamp.UnixNano())
-
-	hasher := sha256.New()
-	hasher.Write([]byte(data))
-	return hex.EncodeToString(hasher.Sum(nil))[:16] // Use first 16 characters
-}
-
-// validateAndSetMetadata validates and sets required metadata for an audit entry
-func (s *GitAuditStore) validateAndSetMetadata(entry *interfaces.AuditEntry) error {
-	// Validate required fields
-	if entry.TenantID == "" {
-		return interfaces.ErrTenantIDRequired
-	}
-	if entry.UserID == "" {
-		return interfaces.ErrUserIDRequired
-	}
-	if entry.Action == "" {
-		return interfaces.ErrActionRequired
-	}
-	if entry.ResourceType == "" {
-		return interfaces.ErrResourceTypeRequired
-	}
-	if entry.ResourceID == "" {
-		return interfaces.ErrResourceIDRequired
-	}
-
-	// Set metadata
-	if entry.ID == "" {
-		entry.ID = s.generateAuditID(entry)
-	}
-	if entry.Timestamp.IsZero() {
-		entry.Timestamp = time.Now()
-	}
-
-	// The audit manager is responsible for checksum calculation
-	// Storage providers should preserve checksums, not generate them
-
-	return nil
-}
-
-// writeAuditFile writes audit entries to a JSON file
-func (s *GitAuditStore) writeAuditFile(filePath string, entries []*interfaces.AuditEntry) error {
-	// Marshal entries to JSON with proper formatting
-	jsonData, err := json.MarshalIndent(entries, "", "  ")
-	if err != nil {
-		return fmt.Errorf("failed to marshal audit entries: %w", err)
-	}
-
-	// Write to file
-	// #nosec G306 - Audit files need read permissions for compliance tools
-	if err := os.WriteFile(filePath, jsonData, 0644); err != nil {
-		return fmt.Errorf("failed to write file: %w", err)
-	}
-
-	return nil
-}
-
-// readAuditFile reads audit entries from a JSON file
-func (s *GitAuditStore) readAuditFile(filePath string) ([]*interfaces.AuditEntry, error) {
-	// Read file contents
-	// #nosec G304 - Git storage requires reading config files from controlled repository paths
-	data, err := os.ReadFile(filePath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to read file: %w", err)
-	}
-
-	// Parse JSON data
-	var entries []*interfaces.AuditEntry
-	if err := json.Unmarshal(data, &entries); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal JSON: %w", err)
-	}
-
-	return entries, nil
-}
-
-// matchesFilter checks if an audit entry matches the filter criteria
-func (s *GitAuditStore) matchesFilter(entry *interfaces.AuditEntry, filter *interfaces.AuditFilter) bool {
-	if filter == nil {
-		return true
-	}
-
-	// Tenant ID filter
-	if filter.TenantID != "" && entry.TenantID != filter.TenantID {
-		return false
-	}
-
-	// Event type filter
-	if len(filter.EventTypes) > 0 {
-		found := false
-		for _, eventType := range filter.EventTypes {
-			if entry.EventType == eventType {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	// Action filter
-	if len(filter.Actions) > 0 {
-		found := false
-		for _, action := range filter.Actions {
-			if entry.Action == action {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	// User ID filter
-	if len(filter.UserIDs) > 0 {
-		found := false
-		for _, userID := range filter.UserIDs {
-			if entry.UserID == userID {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	// User type filter
-	if len(filter.UserTypes) > 0 {
-		found := false
-		for _, userType := range filter.UserTypes {
-			if entry.UserType == userType {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	// Result filter
-	if len(filter.Results) > 0 {
-		found := false
-		for _, result := range filter.Results {
-			if entry.Result == result {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	// Severity filter
-	if len(filter.Severities) > 0 {
-		found := false
-		for _, severity := range filter.Severities {
-			if entry.Severity == severity {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	// Resource type filter
-	if len(filter.ResourceTypes) > 0 {
-		found := false
-		for _, resourceType := range filter.ResourceTypes {
-			if entry.ResourceType == resourceType {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	// Resource ID filter
-	if len(filter.ResourceIDs) > 0 {
-		found := false
-		for _, resourceID := range filter.ResourceIDs {
-			if entry.ResourceID == resourceID {
-				found = true
-				break
-			}
-		}
-		if !found {
-			return false
-		}
-	}
-
-	// Time range filter
-	if filter.TimeRange != nil {
-		if filter.TimeRange.Start != nil && entry.Timestamp.Before(*filter.TimeRange.Start) {
-			return false
-		}
-		if filter.TimeRange.End != nil && entry.Timestamp.After(*filter.TimeRange.End) {
-			return false
-		}
-	}
-
-	// Tags filter
-	if len(filter.Tags) > 0 {
-		entryTags := make(map[string]bool)
-		for _, tag := range entry.Tags {
-			entryTags[tag] = true
-		}
-
-		for _, tag := range filter.Tags {
-			if !entryTags[tag] {
-				return false
-			}
-		}
-	}
-
-	return true
-}
-
-// sortAuditEntries sorts audit entries based on filter criteria
-func (s *GitAuditStore) sortAuditEntries(entries []*interfaces.AuditEntry, filter *interfaces.AuditFilter) {
-	if filter == nil || filter.SortBy == "" {
-		// Default sort by timestamp descending
-		for i, j := 0, len(entries)-1; i < j; i, j = i+1, j-1 {
-			if entries[i].Timestamp.Before(entries[j].Timestamp) {
-				entries[i], entries[j] = entries[j], entries[i]
-			}
-		}
-		return
-	}
-
-	// Custom sorting based on filter criteria
-	// For simplicity, only implement timestamp sorting for now
-	ascending := filter.Order == "asc"
-
-	switch filter.SortBy {
-	case "timestamp":
-		for i, j := 0, len(entries)-1; i < j; i, j = i+1, j-1 {
-			if ascending {
-				if entries[i].Timestamp.After(entries[j].Timestamp) {
-					entries[i], entries[j] = entries[j], entries[i]
-				}
-			} else {
-				if entries[i].Timestamp.Before(entries[j].Timestamp) {
-					entries[i], entries[j] = entries[j], entries[i]
-				}
-			}
-		}
-	}
-}
-
-// gitCommit commits a single file to git for audit operations
-func (s *GitAuditStore) gitCommit(filePath, message string) error {
-	// Make path relative to repository root for git commands
-	relPath, err := filepath.Rel(s.repoPath, filePath)
-	if err != nil {
-		return fmt.Errorf("failed to get relative path: %w", err)
-	}
-
-	// Add file to staging
-	// #nosec G204 - Git audit storage requires controlled git operations
-	cmd := exec.Command("git", "add", relPath)
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to add file to git: %w", err)
-	}
-
-	// Commit the change
-	// #nosec G204 - Git storage requires controlled git command execution
-	cmd = exec.Command("git", "commit", "-m", message)
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to commit to git: %w", err)
-	}
-
-	return nil
-}
-
-// gitCommitMultiple commits multiple files to git in a single commit for audit operations
-func (s *GitAuditStore) gitCommitMultiple(filePaths []string, message string) error {
-	// Make paths relative to repository root for git commands
-	var relPaths []string
-	for _, filePath := range filePaths {
-		relPath, err := filepath.Rel(s.repoPath, filePath)
-		if err != nil {
-			return fmt.Errorf("failed to get relative path: %w", err)
-		}
-		relPaths = append(relPaths, relPath)
-	}
-
-	// Add all files to staging
-	args := append([]string{"add"}, relPaths...)
-	// #nosec G204 - Git storage requires controlled git command execution with validated args
-	cmd := exec.Command("git", args...)
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to add files to git: %w", err)
-	}
-
-	// Commit the changes
-	// #nosec G204 - Git storage requires controlled git command execution
-	cmd = exec.Command("git", "commit", "-m", message)
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to commit to git: %w", err)
-	}
-
-	return nil
-}
-
-// SyncWithRemote synchronizes the audit repository with the remote
-func (s *GitAuditStore) SyncWithRemote(ctx context.Context) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	if s.remoteURL == "" {
-		return nil // No remote configured, skip sync
-	}
-
-	// Pull latest changes from remote
-	if err := s.pullFromRemote(ctx); err != nil {
-		return fmt.Errorf("failed to pull from remote: %w", err)
-	}
-
-	// Push local changes to remote
-	if err := s.pushToRemote(ctx); err != nil {
-		return fmt.Errorf("failed to push to remote: %w", err)
-	}
-
-	return nil
-}
-
-// pullFromRemote pulls changes from the remote repository
-func (s *GitAuditStore) pullFromRemote(ctx context.Context) error {
-	if s.remoteURL == "" {
-		return nil
-	}
-
-	// Check if remote origin exists
-	cmd := exec.CommandContext(ctx, "git", "remote", "get-url", "origin")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		// Add remote origin if it doesn't exist
-		// #nosec G204 - Git storage requires remote URL management for distributed repos
-		cmd = exec.CommandContext(ctx, "git", "remote", "add", "origin", s.remoteURL)
-		cmd.Dir = s.repoPath
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to add remote origin: %w", err)
-		}
-	}
-
-	// Fetch from remote
-	cmd = exec.CommandContext(ctx, "git", "fetch", "origin")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		// Fetch may fail if remote is empty, which is okay for new repos
-		return nil
-	}
-
-	// Pull changes (for audit store, we typically append-only, so conflicts are rare)
-	cmd = exec.CommandContext(ctx, "git", "pull", "origin", "main")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		// Use rebase for append-only audit logs to avoid merge commits
-		cmd = exec.CommandContext(ctx, "git", "pull", "--rebase", "origin", "main")
-		cmd.Dir = s.repoPath
-		if err := cmd.Run(); err != nil {
-			return fmt.Errorf("failed to pull audit data: %w", err)
-		}
-	}
-
-	return nil
-}
-
-// pushToRemote pushes audit changes to the remote repository
-func (s *GitAuditStore) pushToRemote(ctx context.Context) error {
-	if s.remoteURL == "" {
-		return nil
-	}
-
-	// Check if we have any local commits
-	cmd := exec.CommandContext(ctx, "git", "rev-parse", "HEAD")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return nil // No commits to push
-	}
-
-	// Push to remote (audit logs should always be pushed for compliance)
-	cmd = exec.CommandContext(ctx, "git", "push", "origin", "main")
-	cmd.Dir = s.repoPath
-	if err := cmd.Run(); err != nil {
-		return fmt.Errorf("failed to push audit data to remote: %w", err)
-	}
-
-	return nil
-}
-
-// SetRemoteURL sets the remote repository URL for audit storage
-func (s *GitAuditStore) SetRemoteURL(remoteURL string) error {
-	s.mutex.Lock()
-	defer s.mutex.Unlock()
-
-	s.remoteURL = remoteURL
-
-	if remoteURL != "" {
-		// Update or add remote origin
-		cmd := exec.Command("git", "remote", "set-url", "origin", remoteURL)
-		cmd.Dir = s.repoPath
-		if err := cmd.Run(); err != nil {
-			// If set-url fails, try to add the remote
-			cmd = exec.Command("git", "remote", "add", "origin", remoteURL)
-			cmd.Dir = s.repoPath
-			if err := cmd.Run(); err != nil {
-				return fmt.Errorf("failed to set remote URL: %w", err)
-			}
-		}
-	}
-
-	return nil
 }
